navbar completed
commit message after creating tha navbar
there are plenty of problems fixed with storybook and how it rendered different stories created for testing different components and the initial routing system for testing different webpages of the website and how the event handling functions were handling the values received from elements and how the border colors were being rendered conditionally on build time. I also designed the burgermenu icon and the burgermenu itself all from scratch with its animation. the navigation bar at the top of the page was completed in this branch.